refactor: getMappings
