+++ web/netlify/functions/utils/portfolio.ts
@@ -1,73 +1,6 @@
 import { isTwoStringsEqual, isUndefined } from "@/lib/utils";
 import type { SupportedChain } from "@seer-pm/sdk";
-import { COLLATERAL_TOKENS, type Market, getCollateralByIndex, getToken0Token1, getTokensPairKey } from "@seer-pm/sdk";
-import { type Address, erc20Abi } from "viem";
-import { multicall } from "viem/actions";
-import { getPublicClientByChainId } from "./config";
-
-export interface MarketDataMapping {
-  outcomeTokenToCollateral: Map<Address, Address>;
-  conditionIdToMarketMapping: {
-    [key: string]: Market;
-  };
-  allTokensIds: Set<Address>;
-  tokenPairToMarketMapping: {
-    [key: string]: Market;
-  };
-  tokenIdToTokenSymbolMapping: {
-    [key: string]: string;
-  };
-}
-
-export async function getMappings(initialMarkets: Market[], chainId: SupportedChain): Promise<MarketDataMapping> {
-  const markets = initialMarkets.filter((x) => x.chainId === chainId);
-  const conditionIdToMarketMapping: MarketDataMapping["conditionIdToMarketMapping"] = {};
-  const tokenPairToMarketMapping: MarketDataMapping["tokenPairToMarketMapping"] = {};
-  const outcomeTokenToCollateral: MarketDataMapping["outcomeTokenToCollateral"] = new Map();
-  const allTokensIds: MarketDataMapping["allTokensIds"] = new Set();
-
-  for (const market of markets) {
-    conditionIdToMarketMapping[market.conditionId.toLocaleLowerCase()] = market;
-
-    market.wrappedTokens.forEach((outcomeToken, i) => {
-      const collateral = getCollateralByIndex(market, i);
-      const { token0, token1 } = getToken0Token1(collateral, outcomeToken);
-      tokenPairToMarketMapping[getTokensPairKey(token0, token1)] = market;
-
-      outcomeTokenToCollateral.set(outcomeToken.toLocaleLowerCase() as Address, getCollateralByIndex(market, i));
-      allTokensIds.add(market.wrappedTokens[i].toLocaleLowerCase() as Address);
-      allTokensIds.add(collateral.toLocaleLowerCase() as Address);
-    });
-  }
-
-  const publicClient = getPublicClientByChainId(chainId);
-
-  const symbolResults = await multicall(publicClient, {
-    contracts: Array.from(allTokensIds.values()).map((tokenId) => ({
-      abi: erc20Abi,
-      address: tokenId,
-      functionName: "symbol",
-      args: [],
-    })),
-    allowFailure: true,
-  });
-  const allTokensSymbols = symbolResults.map((r) => (r.status === "success" ? (r.result as string) : ""));
-  const tokenIdToTokenSymbolMapping = Array.from(allTokensIds.values()).reduce(
-    (acc, curr, index) => {
-      acc[curr] = allTokensSymbols[index];
-      return acc;
-    },
-    {} as { [key: string]: string },
-  );
-
-  return {
-    outcomeTokenToCollateral,
-    conditionIdToMarketMapping,
-    allTokensIds,
-    tokenPairToMarketMapping,
-    tokenIdToTokenSymbolMapping,
-  };
-}
+import { COLLATERAL_TOKENS } from "@seer-pm/sdk";
 
 export function getTokenPricesMapping(
   positions: { parentMarketId?: string; tokenId: string; collateralToken?: string }[],
